+++ src/components/theme/View/EventView.jsx
@@ -16,10 +16,11 @@ const EventTextfieldView = ({ content }) => (
     {content.description && (
       <p className="documentDescription">{content.description}</p>
     )}
-    {content.image && (
+    {content.preview_image && (
       <Image
         className="document-image"
-        src={content.image.scales.thumb.download}
+        alt={content.preview_caption}
+        src={content.preview_image.scales.thumb.download}
         floated="right"
       />
     )}
@@ -41,13 +42,21 @@ const EventTextfieldView = ({ content }) => (
  */
 const EventView = (props) => {
   const { content } = props;
-
   return (
     <div id="page-document" className="ui container view-wrapper event-view">
       <Grid>
         <Grid.Column width={7} className="mobile hidden">
           {hasBlocksData(content) ? (
-            <RenderBlocks {...props} />
+            <>
+              {content.preview_image && (
+                <Image
+                  className="document-image"
+                  alt={content.preview_caption}
+                  src={content.preview_image.scales.large.download}
+                />
+              )}
+              <RenderBlocks {...props} />
+            </>
           ) : (
             <EventTextfieldView {...props} />
           )}
@@ -58,6 +67,13 @@ const EventView = (props) => {
         <Grid.Column width={12} only="mobile">
           {hasBlocksData(content) ? (
             <>
+              {content.preview_image && (
+                <Image
+                  className="document-image"
+                  alt={content.preview_caption}
+                  src={content.preview_image.scales.mini.download}
+                />
+              )}
               <RenderBlocks
                 {...props}
                 content={{